# backend/openapi.internal.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SaaS Boilerplate",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/billing/plans": {
      "get": {
        "tags": [
          "Billing Plans"
        ],
        "summary": "List Plans",
        "operationId": "list_plans_v1_billing_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanOut"
                  },
                  "type": "array",
                  "title": "Response List Plans V1 Billing Plans Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/billing/plans/{plan_id}": {
      "get": {
        "tags": [
          "Billing Plans"
        ],
        "summary": "Retrieve A Plan",
        "operationId": "retrieve_a_plan_v1_billing_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/billing/subscriptions/checkout": {
      "post": {
        "tags": [
          "Billing Subscriptions"
        ],
        "summary": "Start Checkout",
        "operationId": "start_checkout_v1_billing_subscriptions_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/billing/subscriptions/trial": {
      "post": {
        "tags": [
          "Billing Subscriptions"
        ],
        "summary": "Start Trial",
        "operationId": "start_trial_v1_billing_subscriptions_trial_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTrialIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/billing/subscriptions/current": {
      "get": {
        "tags": [
          "Billing Subscriptions"
        ],
        "summary": "Get Current Subscription",
        "operationId": "get_current_subscription_v1_billing_subscriptions_current_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Billing Subscriptions"
        ],
        "summary": "Update Billing Email",
        "operationId": "update_billing_email_v1_billing_subscriptions_current_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBillingEmailIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/billing/subscriptions/current/cancel": {
      "post": {
        "tags": [
          "Billing Subscriptions"
        ],
        "summary": "Cancel Subscription",
        "operationId": "cancel_subscription_v1_billing_subscriptions_current_cancel_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/billing/subscriptions/current/resume": {
      "post": {
        "tags": [
          "Billing Subscriptions"
        ],
        "summary": "Resume Subscription",
        "operationId": "resume_subscription_v1_billing_subscriptions_current_resume_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/billing/subscriptions/current/switch-plan": {
      "post": {
        "tags": [
          "Billing Subscriptions"
        ],
        "summary": "Switch Plan",
        "operationId": "switch_plan_v1_billing_subscriptions_current_switch_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchPlanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/billing/subscriptions/portal": {
      "get": {
        "tags": [
          "Billing Subscriptions"
        ],
        "summary": "Get Customer Portal",
        "operationId": "get_customer_portal_v1_billing_subscriptions_portal_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPortalOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/billing/usage": {
      "get": {
        "tags": [
          "Billing Usage"
        ],
        "summary": "Get Current Usage",
        "operationId": "get_current_usage_v1_billing_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/billing/webhook": {
      "post": {
        "tags": [
          "Billing Webhooks"
        ],
        "summary": "Billing Webhook",
        "operationId": "billing_webhook_v1_billing_webhook_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Billing Webhook V1 Billing Webhook Post"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutIn": {
        "properties": {
          "plan_price_id": {
            "type": "integer",
            "title": "Plan Price Id"
          }
        },
        "type": "object",
        "required": [
          "plan_price_id"
        ],
        "title": "CheckoutIn"
      },
      "CheckoutOut": {
        "properties": {
          "checkout_url": {
            "type": "string",
            "title": "Checkout Url"
          }
        },
        "type": "object",
        "required": [
          "checkout_url"
        ],
        "title": "CheckoutOut"
      },
...
      "CustomerPortalOut": {
        "properties": {
          "portal_url": {
            "type": "string",
            "title": "Portal Url"
          }
        },
        "type": "object",
        "required": [
          "portal_url"
        ],
        "title": "CustomerPortalOut"
      },
...
      "PlanOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/PlanPriceOut"
            },
            "type": "array",
            "title": "Prices"
          },
          "plan_settings": {
            "items": {
              "$ref": "#/components/schemas/PlanSettingOut"
            },
            "type": "array",
            "title": "Plan Settings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "is_active",
          "prices",
          "plan_settings",
          "created_at",
          "updated_at"
        ],
        "title": "PlanOut"
      },
      "PlanPriceOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "interval": {
            "type": "string",
            "title": "Interval"
          },
          "interval_count": {
            "type": "integer",
            "title": "Interval Count"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "trial_period_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Period Days",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_id",
          "amount",
          "currency",
          "interval",
          "interval_count",
          "is_active",
          "created_at",
          "updated_at",
          "trial_period_days"
        ],
        "title": "PlanPriceOut"
      },
      "PlanSettingOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "PlanSettingOut"
      },
...
      "StartTrialIn": {
        "properties": {
          "plan_price_id": {
            "type": "integer",
            "title": "Plan Price Id"
          }
        },
        "type": "object",
        "required": [
          "plan_price_id"
        ],
        "title": "StartTrialIn"
      },
      "SubscriptionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "plan_price_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Price Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_period_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period Start"
          },
          "current_period_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "cancel_at_period_end": {
            "type": "boolean",
            "title": "Cancel At Period End"
          },
          "canceled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canceled At"
          },
          "cancel_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel At"
          },
          "trial_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial End"
          },
          "plan_price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanPriceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "billing_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Email"
          },
          "has_payment_method": {
            "type": "boolean",
            "title": "Has Payment Method",
            "default": false
          },
          "trial_used": {
            "type": "boolean",
            "title": "Trial Used",
            "default": false
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "plan_settings": {
            "items": {
              "$ref": "#/components/schemas/PlanSettingOut"
            },
            "type": "array",
            "title": "Plan Settings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "has_access": {
            "type": "boolean",
            "title": "Has Access",
            "description": "Whether the plan applies (`ENTITLED_STATUSES`) - the app's access\npolicy, so the frontend never keeps its own list of statuses.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "plan_price_id",
          "status",
          "current_period_start",
          "current_period_end",
          "cancel_at_period_end",
          "canceled_at",
          "cancel_at",
          "trial_end",
          "plan_price",
          "billing_email",
          "has_payment_method",
          "trial_used",
          "features",
          "plan_settings",
          "created_at",
          "updated_at",
          "has_access"
        ],
        "title": "SubscriptionOut"
      },
...
      "SwitchPlanIn": {
        "properties": {
          "plan_price_id": {
            "type": "integer",
            "title": "Plan Price Id"
          }
        },
        "type": "object",
        "required": [
          "plan_price_id"
        ],
        "title": "SwitchPlanIn"
      },
...
      "UpdateBillingEmailIn": {
        "properties": {
          "billing_email": {
            "type": "string",
            "format": "email",
            "title": "Billing Email"
          }
        },
        "type": "object",
        "required": [
          "billing_email"
        ],
        "title": "UpdateBillingEmailIn"
      },
      "UsageItemOut": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "metric",
          "count",
          "limit"
        ],
        "title": "UsageItemOut"
      },
      "UsageOut": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "usage": {
            "items": {
              "$ref": "#/components/schemas/UsageItemOut"
            },
            "type": "array",
            "title": "Usage"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "usage"
        ],
        "title": "UsageOut"
      },
...
      "ErrorResponse": {
        "properties": {
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "time",
          "path",
          "method",
          "error_code",
          "msg"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ValidationErrorResponse": {
        "$defs": {
          "ValidationDetail": {
            "properties": {
              "error_code": {
                "title": "Error Code",
                "type": "string"
              },
              "field": {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                },
                "title": "Field",
                "type": "array"
              },
              "msg": {
                "title": "Msg",
                "type": "string"
              },
              "ctx": {
                "additionalProperties": true,
                "title": "Ctx",
                "type": "object"
              }
            },
            "required": [
              "error_code",
              "field",
              "msg",
              "ctx"
            ],
            "title": "ValidationDetail",
            "type": "object"
          }
        },
        "properties": {
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationDetail"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "time",
          "path",
          "method",
          "error_code",
          "msg",
          "errors"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "v1/auth/token"
          }
        }
      }
    }
  }
}